Trimmed large svelte files; added tests, README update
