+++ .github/workflows/ci.yml
@@ -1,0 +1,33 @@
+name: CI Check
+
+on:
+  push:
+    branches: [ Dev, master ]
+  pull_request:
+    branches: [ Dev, master ]
+
+jobs:
+  build-and-lint:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout Code
+        uses: actions/checkout@v4
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20
+          cache: 'npm'
+
+      - name: Install Dependencies
+        run: npm ci
+
+      - name: Check TypeScript & Build Bundle
+        run: npm run build
+
+      - name: Run ESLint
+        run: npm run lint
+
+      - name: Run Unit Tests
+        run: npm test
