+++ package.json
@@ -7,7 +7,8 @@
 		"dev": "node esbuild.config.mjs",
 		"build": "tsc -noEmit -skipLibCheck && node esbuild.config.mjs production",
 		"version": "node version-bump.mjs && git add manifest.json versions.json docs-site/versioned_docs docs-site/versioned_sidebars docs-site/versions.json",
-		"lint": "eslint ."
+		"lint": "eslint .",
+		"test": "vitest run"
 	},
 	"keywords": [
 		"obsidian",
@@ -29,7 +30,7 @@
 		"@eslint/js": "^9.0.0",
 		"@eslint/json": "^2.0.0",
 		"@lezer/highlight": "^1.2.3",
-		"@types/node": "^16.18.126",
+		"@types/node": "^20.0.0",
 		"@typescript-eslint/parser": "^8.61.0",
 		"esbuild": "0.17.3",
 		"esbuild-svelte": "^0.9.3",
@@ -41,7 +42,8 @@
 		"ts-node": "^10.9.2",
 		"tslib": "2.4.0",
 		"typescript": "^5.9.3",
-		"typescript-eslint": "^8.61.0"
+		"typescript-eslint": "^8.61.0",
+		"vitest": "^1.6.0"
 	},
 	"dependencies": {
 		"chart.js": "^4.5.1",
