+++ vitest.config.ts
@@ -1,0 +1,13 @@
+import { defineConfig } from 'vitest/config';
+import path from 'path';
+
+export default defineConfig({
+	test: {
+		globals: true,
+		environment: 'node',
+		include: ['tests/**/*.test.ts'],
+		alias: {
+			obsidian: path.resolve(__dirname, 'tests/mocks/obsidian.ts'),
+		},
+	},
+});
